+++ packages/frontend/src/downstream/Context/Output/Character/Wrapper.tsx
@@ -3,18 +3,26 @@ import styled from "styled-components";
 
 import { Wrapper as Animation } from "./Animation/Wrapper";
 import { Wrapper as Quiz } from "./Quiz/Wrapper";
+import { Title } from "./Title";
 
 interface IWrapper {
   text: string;
   stroke: any;
 }
 
-const WrapperView = styled("div")``;
+const WrapperView = styled("div")`
+  display: flex;
+  align-items: center;
+  & > *:nth-child(2) {
+    margin-left: 1em;
+  }
+`;
 
 const Wrapper = ({ text, stroke, ...otherProps }: IWrapper) => (
   <WrapperView {...otherProps}>
-    <Animation text={text} stroke={stroke} />
-    <Quiz text={text} stroke={stroke} />
+    <Title title={text} />
+    <Animation stroke={stroke} />
+    <Quiz stroke={stroke} />
   </WrapperView>
 );